+++ features/support/pages/legal_panel_for_government_page.rb
@@ -14,5 +14,10 @@ module Pages
       element :yes, '#have_you_reviewed_yes'
       element :no, '#have_you_reviewed_no'
     end
+
+    section :questions, '#main-content' do
+      element :'requirement replace an existing contract', '#replaces_existing_contract-form-group'
+      element :'CCS contact you', '#ccs_can_contact_you-form-group'
+    end
   end
 end