+++ spec/models/legal_panel_for_government/rm6360/journey/information_about_your_requirement_spec.rb
@@ -162,13 +162,13 @@ RSpec.describe LegalPanelForGovernment::RM6360::Journey::InformationAboutYourReq
 
   describe '.permit_list' do
     it 'returns a list of the permitted attributes' do
-      expect(described_class.permit_list).to eq [:requirement_start_date_day, :requirement_start_date_month, :requirement_start_date_year, :requirement_end_date_day, :requirement_end_date_month, :requirement_end_date_year, :requirement_estimated_total_value, :ccs_can_contact_you, :replaces_existing_contract, {}]
+      expect(described_class.permit_list).to eq [:requirement_start_date_day, :requirement_start_date_month, :requirement_start_date_year, :requirement_end_date_day, :requirement_end_date_month, :requirement_end_date_year, :requirement_estimated_total_value, :replaces_existing_contract, :ccs_can_contact_you, {}]
     end
   end
 
   describe '.permitted_keys' do
     it 'returns a list of the permitted keys' do
-      expect(described_class.permitted_keys).to eq %i[requirement_start_date_day requirement_start_date_month requirement_start_date_year requirement_end_date_day requirement_end_date_month requirement_end_date_year requirement_estimated_total_value ccs_can_contact_you replaces_existing_contract]
+      expect(described_class.permitted_keys).to eq %i[requirement_start_date_day requirement_start_date_month requirement_start_date_year requirement_end_date_day requirement_end_date_month requirement_end_date_year requirement_estimated_total_value replaces_existing_contract ccs_can_contact_you]
     end
   end