+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add flag to show GCA branding ([PR 2273](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/2273))
+
 ### Changed
 
 - Update GOV.UK Frontend and CCS Frontend ([PR 2251](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2251))